+++ sdk/src/main/java/com/mapbox/maps/coroutine/MapboxMapExt.kt
@@ -10,8 +10,10 @@ import com.mapbox.bindgen.Expected
 import com.mapbox.bindgen.None
 import com.mapbox.bindgen.Value
 import com.mapbox.geojson.Feature
+import com.mapbox.geojson.Point
 import com.mapbox.maps.*
 import com.mapbox.maps.extension.style.StyleContract
+import com.mapbox.maps.util.isEmpty
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.channels.trySendBlocking
@@ -448,4 +450,34 @@ fun MapboxMap.genericEvents(eventName: String): Flow<GenericEvent> {
       nativeObserver.subscribeGenericEvent(eventName, ::trySendBlocking, onCancel = channel::close)
     awaitClose(cancelable::cancel)
   }.flowOn(Dispatchers.Main.immediate)
+}
+
+/**
+ * Convenience method that returns the [CameraOptions] object for given parameters.
+ *
+ * @param coordinates The `coordinates` representing the bounds of the camera.
+ * @param camera The [CameraOptions] which will be applied before calculating the camera for the coordinates. If any of the fields in [CameraOptions] are not provided then the current value from the map for that field will be used.
+ * @param coordinatesPadding The amount of padding in pixels to add to the given `coordinates`.
+ *                           Note: This padding is not applied to the map but to the coordinates provided. If you want to apply padding to the map use param `camera`.
+ * @param maxZoom The maximum zoom level allowed in the returned camera options.
+ * @param offset The center of the given bounds relative to map center in pixels.
+ *
+ * @return the [CameraOptions] object representing the provided parameters. Empty [CameraOptions] (see [CameraOptions.isEmpty]) could be returned only if an internal error occurred.
+ */
+@JvmSynthetic
+suspend fun MapboxMap.cameraForCoordinates(
+  coordinates: List<Point>,
+  camera: CameraOptions,
+  coordinatesPadding: EdgeInsets? = null,
+  maxZoom: Double? = null,
+  offset: ScreenCoordinate? = null,
+): CameraOptions = suspendCoroutine { continuation ->
+  cameraForCoordinates(
+    coordinates = coordinates,
+    camera = camera,
+    coordinatesPadding = coordinatesPadding,
+    maxZoom = maxZoom,
+    offset = offset,
+    result = continuation::resume
+  )
 }